+++ contracts/freeportParts/ERC20Adapter.sol
@@ -45,4 +45,9 @@ abstract contract ERC20Adapter is TransferOperator {
         currencyContract.transfer(user, amount);
     }
 
+    function sendShare(address beneficiar, uint256 amount) internal {
+        _burn(beneficiar, CURRENCY, amount);
+        currencyContract.transfer(beneficiar, amount);
+    }
+
 }